controllers/EDGE-MQTT-AZURETABLES/requirements.txt
#azure-eventhub
azure-data-tables
termcolor
paho-mqtt